sphinx autodoc was unable to find class in __init__ file, so moved it to new file
