+++ docs/api/calibration/base.rst
@@ -1,0 +1,12 @@
+``columnflow.calibration.base``
+===============================
+.. automodule:: columnflow.calibration.base
+.. currentmodule:: columnflow.calibration.base
+   
+Summary
+-------
+
+..   autoclass:: Calibrator
+    :members:
+
+    :inherit-members:
